Include :provided profile when compiling
The :provided profile designates dependencies that are to be excluded
from uberjars.  This change also implements the exclusion for uberwars.
It is compatible with the horrible logic carried over from lein-ring
that excludes any jars from the uberjar that provide the servlet
interface, but it allows for additional jars to be excluded, e.g.:

```
:profiles {:provided {:dependencies [[javax.servlet/javax.servlet-api "3.1.0"]
                                     [log4j/log4j "1.2.17"]
                                     [org.camunda.bpm/camunda-engine "7.4.0"]]}}
```

Would make the listed dependencies available when compiling for the .war
file, but not include them in the uberjar.